Refactor, as per CodeRabbit feedback
Reject as soon as activateMainMenuItem rejects (if it does)
