+++ packages/app-desktop/integration-tests/models/MainScreen.ts
@@ -7,7 +7,7 @@ import setFilePickerResponse from '../util/setFilePickerResponse';
 import NoteList from './NoteList';
 import { expect } from '../util/test';
 import ChangeAppLayoutScreen from './ChangeAppLayoutScreen';
-import { Second } from '@joplin/utils/time';
+import waitForNextWindowMatching from '../util/waitForNextWindowMatching';
 
 export default class MainScreen {
 	public readonly newNoteButton: Locator;
@@ -66,35 +66,10 @@ export default class MainScreen {
 	}
 
 	public async openNewWindow(electronApp: ElectronApplication) {
-		return new Promise<Page>((resolve, reject) => {
-			let timeout: NodeJS.Timeout|null = null;
-			const clearListenersAndTimeouts = () => {
-				if (timeout) {
-					clearTimeout(timeout);
-					timeout = null;
-				}
-				electronApp.off('window', onWindowAdded);
-			};
+		const pagePromise = waitForNextWindowMatching(/^Joplin - /, electronApp);
 
-			const onWindowAdded = async (page: Page) => {
-				if ((await page.title()).startsWith('Joplin -')) {
-					clearListenersAndTimeouts();
-					resolve(page);
-				}
-			};
-			electronApp.on('window', onWindowAdded);
-
-			timeout = setTimeout(async () => {
-				timeout = null;
-				clearListenersAndTimeouts();
-
-				const windows = electronApp.windows();
-				const titles = await Promise.all(windows.map(w => w.title()));
-				reject(new Error(`Opening a window timed out. Open window titles: ${JSON.stringify(titles)}.`));
-			}, 30 * Second);
-
-			void activateMainMenuItem(electronApp, 'Open in new window');
-		});
+		await activateMainMenuItem(electronApp, 'Open in new window');
+		return pagePromise;
 	}
 
 	public async search(text: string) {
